+++ packages/wasm/crate/tests/utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod planner_setup;
+pub mod sct;
